renaming the component to make more generic
