+++ core-web/libs/template-builder/src/lib/components/template-builder/components/remove-confirm-dialog/remove-confirm-dialog.component.html
@@ -1,7 +1,7 @@
 <p-confirmPopup></p-confirmPopup>
 <p-button
     (onClick)="confirm($event)"
-    testId="btn-remove-row"
+    data-testId="btn-remove-item"
     styleClass="p-button-text p-button-sm p-button-rounded"
     icon="pi pi-trash"
 >
