+++ core-web/libs/template-builder/src/lib/components/template-builder/components/remove-confirm-dialog/remove-confirm-dialog.component.scss
@@ -17,4 +17,8 @@
     .p-confirm-popup .p-confirm-popup-content {
         padding: $spacing-3;
     }
+
+    .p-button.p-button-text .pi.pi-trash {
+        color: $color-palette-primary-400;
+    }
 }
