+++ core-web/libs/template-builder/src/lib/components/template-builder/components/template-builder-row/template-builder-row.component.html
@@ -18,5 +18,5 @@
         icon="pi pi-palette"
     >
     </p-button>
-    <dotcms-remove-row (deleteRow)="deleteRow.emit()"></dotcms-remove-row>
+    <dotcms-remove-confirm-dialog (confirmEvent)="deleteRow.emit()"></dotcms-remove-confirm-dialog>
 </div>
