+++ core-web/libs/template-builder/src/lib/components/template-builder/components/template-builder-row/template-builder-row.component.scss
@@ -33,11 +33,3 @@
     flex-grow: 1;
     padding: $spacing-3 0;
 }
-
-.row__actions-container {
-    p-button ::ng-deep {
-        .pi-trash {
-            color: $color-palette-primary-400;
-        }
-    }
-}
