+++ core-web/libs/template-builder/src/lib/components/template-builder/template-builder.component.stories.ts
@@ -4,7 +4,7 @@ import { NgFor, AsyncPipe } from '@angular/common';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
 import { AddWidgetComponent } from './components/add-widget/add-widget.component';
-import { RemoveRowComponent } from './components/remove-row/remove-row.component';
+import { RemoveConfirmDialogComponent } from './components/remove-confirm-dialog/remove-confirm-dialog.component';
 import { TemplateBuilderRowComponent } from './components/template-builder-row/template-builder-row.component';
 import { DotTemplateBuilderStore } from './store/template-builder.store';
 import { TemplateBuilderComponent } from './template-builder.component';
@@ -20,7 +20,7 @@ export default {
                 AsyncPipe,
                 TemplateBuilderRowComponent,
                 AddWidgetComponent,
-                RemoveRowComponent,
+                RemoveConfirmDialogComponent,
                 BrowserAnimationsModule
             ],
             providers: [DotTemplateBuilderStore]
