+++ core-web/libs/template-builder/src/lib/template-builder.module.ts
@@ -2,7 +2,7 @@ import { AsyncPipe, NgFor } from '@angular/common';
 import { NgModule } from '@angular/core';
 
 import { AddWidgetComponent } from './components/template-builder/components/add-widget/add-widget.component';
-import { RemoveRowComponent } from './components/template-builder/components/remove-row/remove-row.component';
+import { RemoveConfirmDialogComponent } from './components/template-builder/components/remove-confirm-dialog/remove-confirm-dialog.component';
 import { TemplateBuilderBoxComponent } from './components/template-builder/components/template-builder-box/template-builder-box.component';
 import { TemplateBuilderRowComponent } from './components/template-builder/components/template-builder-row/template-builder-row.component';
 import { DotTemplateBuilderStore } from './components/template-builder/store/template-builder.store';
@@ -12,7 +12,7 @@ import { TemplateBuilderComponent } from './components/template-builder/template
     imports: [
         NgFor,
         AsyncPipe,
-        RemoveRowComponent,
+        RemoveConfirmDialogComponent,
         TemplateBuilderRowComponent,
         AddWidgetComponent,
         TemplateBuilderBoxComponent
